+++ src/Features.md
@@ -1,11 +1,13 @@
-1- Ask user if words are proper nouns and change if so 
+1- Change sh to siy
 
-2- Ask user if c sounds like s or k
+2- Change ph to f
 
-3- Ask user if ch sounds like ch or k
+3- Capitalize vowels if there are more than one
 
-4- Change sh to siy
+3- Ask user if words are proper nouns and change if so
 
-5- Change ph to f
+4- Ask user if c sounds like s or k
 
-5- Update styles
+5- Ask user if ch sounds like ch or k
+
+6- Update styles
