+++ src/app/services/baybayin/replacement-logic/rules/add-plus-after-consonant.service.spec.ts
