+++ src/app/services/baybayin/replacement-logic/rules/capitalize-first-vowel.service.ts
