+++ src/app/services/baybayin/replacement-logic/rules/convert-ng-and-mga.service.spec.ts
