+++ src/app/services/baybayin/replacement-logic/rules/remove-a-after-consonant.service.ts
