+++ src/app/services/baybayin/replacement-logic/rules/replace-c-with-k.service.ts
