+++ src/app/services/baybayin/replacement-logic/rules/replace-ng-with-capital-n.service.ts
