+++ src/app/services/baybayin/replacement-logic/rules/replace-sh-with-siy.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ReplaceShWithSiyService } from './replace-sh-with-siy.service';
+
+describe('ReplaceShWithSiyService', () => {
+  let service: ReplaceShWithSiyService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ReplaceShWithSiyService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
